third_party/WebKit/LayoutTests/platform/mac-mac10.11/virtual/outofblink-cors-http/http/tests/fetch/serviceworker/thorough/cors-preflight2-other-https-expected.txt

This is a testharness.js-based test.
PASS Startup 
PASS cors-preflight2 
PASS executeTest-0 
PASS executeTest-1 
PASS executeTest-2 
FAIL executeTest-3 assert_equals: https://localhost:8443/serviceworker/resources/fetch-access-control.php?mode=cors&credentials=same-origin&method=PUT&ACAOrigin=*&ACAMethods=PUT should be resolved expected "resolved" but got "rejected"
PASS executeTest-4 
FAIL executeTest-5 assert_equals: https://localhost:8443/serviceworker/resources/fetch-access-control.php?mode=cors&credentials=same-origin&method=PUT&ACAOrigin=*&headers=CUSTOM&ACAMethods=PUT&ACAHeaders=x-serviceworker-test should be resolved expected "resolved" but got "rejected"
FAIL executeTest-6 assert_equals: https://localhost:8443/serviceworker/resources/fetch-access-control.php?mode=cors&credentials=same-origin&method=PUT&ACAOrigin=*&headers=CUSTOM&ACAMethods=PUT&ACAHeaders=x-serviceworker-test&ACEHeaders=Content-Length, X-ServiceWorker-ServerHeader should be resolved expected "resolved" but got "rejected"
PASS executeTest-7 
FAIL executeTest-8 assert_equals: https://localhost:8443/serviceworker/resources/fetch-access-control.php?mode=cors&credentials=same-origin&method=PUT&ACAOrigin=*&headers=CUSTOM&ACAMethods=PUT, XXX&ACAHeaders=x-serviceworker-test should be resolved expected "resolved" but got "rejected"
FAIL executeTest-9 assert_equals: https://localhost:8443/serviceworker/resources/fetch-access-control.php?mode=cors&credentials=same-origin&method=PUT&ACAOrigin=*&headers=CUSTOM&ACAMethods=PUT, XXX&ACAHeaders=x-serviceworker-test&ACEHeaders=Content-Length, X-ServiceWorker-ServerHeader should be resolved expected "resolved" but got "rejected"
PASS executeTest-10 
FAIL executeTest-11 assert_equals: https://localhost:8443/serviceworker/resources/fetch-access-control.php?mode=cors&credentials=same-origin&method=PUT&ACAOrigin=http://127.0.0.1:8000&ACAMethods=PUT should be resolved expected "resolved" but got "rejected"
PASS executeTest-12 
FAIL executeTest-13 assert_equals: https://localhost:8443/serviceworker/resources/fetch-access-control.php?mode=cors&credentials=same-origin&method=PUT&ACAOrigin=http://127.0.0.1:8000&headers=CUSTOM&ACAMethods=PUT&ACAHeaders=x-serviceworker-test should be resolved expected "resolved" but got "rejected"
FAIL executeTest-14 assert_equals: https://localhost:8443/serviceworker/resources/fetch-access-control.php?mode=cors&credentials=same-origin&method=PUT&ACAOrigin=http://127.0.0.1:8000&headers=CUSTOM&ACAMethods=PUT&ACAHeaders=x-serviceworker-test&ACEHeaders=Content-Length, X-ServiceWorker-ServerHeader should be resolved expected "resolved" but got "rejected"
PASS executeTest-15 
FAIL executeTest-16 assert_equals: https://localhost:8443/serviceworker/resources/fetch-access-control.php?mode=cors&credentials=same-origin&method=PUT&ACAOrigin=http://127.0.0.1:8000&headers=CUSTOM&ACAMethods=PUT, XXX&ACAHeaders=x-serviceworker-test should be resolved expected "resolved" but got "rejected"
FAIL executeTest-17 assert_equals: https://localhost:8443/serviceworker/resources/fetch-access-control.php?mode=cors&credentials=same-origin&method=PUT&ACAOrigin=http://127.0.0.1:8000&headers=CUSTOM&ACAMethods=PUT, XXX&ACAHeaders=x-serviceworker-test&ACEHeaders=Content-Length, X-ServiceWorker-ServerHeader should be resolved expected "resolved" but got "rejected"
FAIL executeTest-18 assert_equals: https://localhost:8443/serviceworker/resources/fetch-access-control.php?mode=cors&credentials=same-origin&method=PUT&ACAOrigin=*&PACAOrigin=*&PACAMethods=PUT&PreflightTest=200 should be resolved expected "resolved" but got "rejected"
PASS executeTest-19 
FAIL executeTest-20 assert_equals: https://localhost:8443/serviceworker/resources/fetch-access-control.php?mode=cors&credentials=same-origin&method=PUT&headers=CUSTOM&ACAOrigin=*&PACAOrigin=*&PACAMethods=PUT&PACAHeaders=x-serviceworker-test&PreflightTest=200 should be resolved expected "resolved" but got "rejected"
PASS executeTest-21 
PASS executeTest-22 
PASS executeTest-23 
FAIL executeTest-24 assert_equals: https://localhost:8443/serviceworker/resources/fetch-access-control.php?mode=cors&credentials=same-origin&method=PUT&ACAOrigin=*&PACAOrigin=*&PACAMethods=PUT&PACEHeaders=Content-Length, X-ServiceWorker-ServerHeader&PreflightTest=200 should be resolved expected "resolved" but got "rejected"
FAIL executeTest-25 assert_equals: https://localhost:8443/serviceworker/resources/fetch-access-control.php?mode=cors&credentials=same-origin&method=PUT&ACAOrigin=*&PACAOrigin=*&PACAMethods=PUT&PreflightTest=201 should be resolved expected "resolved" but got "rejected"
PASS executeTest-26 
PASS executeTest-27 
PASS executeTest-28 
FAIL executeTest-29 assert_equals: https://localhost:8443/serviceworker/resources/fetch-access-control.php?mode=cors&credentials=same-origin&method=PUT&headers=CUSTOM2&ACAOrigin=*&PACAOrigin=*&PACAMethods=PUT&PACAHeaders=x-servicEworker-u,x-servicEworker-ua,x-servicewOrker-test,x-sErviceworker-s,x-sErviceworker-v&PreflightTest=200 should be resolved expected "resolved" but got "rejected"
PASS executeTest-30 
FAIL executeTest-31 assert_equals: https://localhost:8443/serviceworker/resources/fetch-access-control.php?mode=cors&credentials=same-origin&method=PUT&headers=CUSTOM&ACAOrigin=*&PACAOrigin=*&PACAMethods=PUT&PACAHeaders=x-serviceworker-test&PACRMethod=PUT&PACRHeaders=x-serviceworker-test&PreflightTest=200 should be resolved expected "resolved" but got "rejected"
FAIL executeTest-32 assert_equals: https://localhost:8443/serviceworker/resources/fetch-access-control.php?mode=cors&credentials=same-origin&method=PUT&headers=SAFE&ACAOrigin=*&PACAOrigin=*&PACAMethods=PUT&PACRHeaders=missing&PACRMethod=PUT&PreflightTest=200 should be resolved expected "resolved" but got "rejected"
FAIL executeTest-33 assert_equals: https://localhost:8443/serviceworker/resources/fetch-access-control.php?mode=cors&credentials=same-origin&method=PUT&headers=CUSTOM2&ACAOrigin=*&PACAOrigin=*&PACAMethods=PUT&PACAHeaders=x-servicEworker-u,x-servicEworker-ua,x-servicewOrker-test,x-sErviceworker-s,x-sErviceworker-v&PACRMethod=PUT&PACRHeaders=x-serviceworker-s,x-serviceworker-test,x-serviceworker-u,x-serviceworker-ua,x-serviceworker-v&PreflightTest=200 should be resolved expected "resolved" but got "rejected"
PASS executeTest-34 
PASS executeTest-35 
PASS executeTest-36 
FAIL executeTest-37 assert_equals: https://localhost:8443/serviceworker/resources/fetch-access-control.php?mode=cors&credentials=same-origin&method=XXX&ACAOrigin=*&ACAMethods=XXX should be resolved expected "resolved" but got "rejected"
PASS executeTest-38 
FAIL executeTest-39 assert_equals: https://localhost:8443/serviceworker/resources/fetch-access-control.php?mode=cors&credentials=same-origin&method=XXX&ACAOrigin=*&headers=CUSTOM&ACAMethods=XXX&ACAHeaders=x-serviceworker-test should be resolved expected "resolved" but got "rejected"
FAIL executeTest-40 assert_equals: https://localhost:8443/serviceworker/resources/fetch-access-control.php?mode=cors&credentials=same-origin&method=XXX&ACAOrigin=*&headers=CUSTOM&ACAMethods=XXX&ACAHeaders=x-serviceworker-test&ACEHeaders=Content-Length, X-ServiceWorker-ServerHeader should be resolved expected "resolved" but got "rejected"
PASS executeTest-41 
FAIL executeTest-42 assert_equals: https://localhost:8443/serviceworker/resources/fetch-access-control.php?mode=cors&credentials=same-origin&method=XXX&ACAOrigin=*&headers=CUSTOM&ACAMethods=PUT, XXX&ACAHeaders=x-serviceworker-test should be resolved expected "resolved" but got "rejected"
FAIL executeTest-43 assert_equals: https://localhost:8443/serviceworker/resources/fetch-access-control.php?mode=cors&credentials=same-origin&method=XXX&ACAOrigin=*&headers=CUSTOM&ACAMethods=PUT, XXX&ACAHeaders=x-serviceworker-test&ACEHeaders=Content-Length, X-ServiceWorker-ServerHeader should be resolved expected "resolved" but got "rejected"
PASS executeTest-44 
FAIL executeTest-45 assert_equals: https://localhost:8443/serviceworker/resources/fetch-access-control.php?mode=cors&credentials=same-origin&method=XXX&ACAOrigin=http://127.0.0.1:8000&ACAMethods=XXX should be resolved expected "resolved" but got "rejected"
PASS executeTest-46 
FAIL executeTest-47 assert_equals: https://localhost:8443/serviceworker/resources/fetch-access-control.php?mode=cors&credentials=same-origin&method=XXX&ACAOrigin=http://127.0.0.1:8000&headers=CUSTOM&ACAMethods=XXX&ACAHeaders=x-serviceworker-test should be resolved expected "resolved" but got "rejected"
FAIL executeTest-48 assert_equals: https://localhost:8443/serviceworker/resources/fetch-access-control.php?mode=cors&credentials=same-origin&method=XXX&ACAOrigin=http://127.0.0.1:8000&headers=CUSTOM&ACAMethods=XXX&ACAHeaders=x-serviceworker-test&ACEHeaders=Content-Length, X-ServiceWorker-ServerHeader should be resolved expected "resolved" but got "rejected"
PASS executeTest-49 
FAIL executeTest-50 assert_equals: https://localhost:8443/serviceworker/resources/fetch-access-control.php?mode=cors&credentials=same-origin&method=XXX&ACAOrigin=http://127.0.0.1:8000&headers=CUSTOM&ACAMethods=PUT, XXX&ACAHeaders=x-serviceworker-test should be resolved expected "resolved" but got "rejected"
FAIL executeTest-51 assert_equals: https://localhost:8443/serviceworker/resources/fetch-access-control.php?mode=cors&credentials=same-origin&method=XXX&ACAOrigin=http://127.0.0.1:8000&headers=CUSTOM&ACAMethods=PUT, XXX&ACAHeaders=x-serviceworker-test&ACEHeaders=Content-Length, X-ServiceWorker-ServerHeader should be resolved expected "resolved" but got "rejected"
FAIL executeTest-52 assert_equals: https://localhost:8443/serviceworker/resources/fetch-access-control.php?mode=cors&credentials=same-origin&method=XXX&ACAOrigin=*&PACAOrigin=*&PACAMethods=XXX&PreflightTest=200 should be resolved expected "resolved" but got "rejected"
PASS executeTest-53 
FAIL executeTest-54 assert_equals: https://localhost:8443/serviceworker/resources/fetch-access-control.php?mode=cors&credentials=same-origin&method=XXX&headers=CUSTOM&ACAOrigin=*&PACAOrigin=*&PACAMethods=XXX&PACAHeaders=x-serviceworker-test&PreflightTest=200 should be resolved expected "resolved" but got "rejected"
PASS executeTest-55 
PASS executeTest-56 
PASS executeTest-57 
FAIL executeTest-58 assert_equals: https://localhost:8443/serviceworker/resources/fetch-access-control.php?mode=cors&credentials=same-origin&method=XXX&ACAOrigin=*&PACAOrigin=*&PACAMethods=XXX&PACEHeaders=Content-Length, X-ServiceWorker-ServerHeader&PreflightTest=200 should be resolved expected "resolved" but got "rejected"
FAIL executeTest-59 assert_equals: https://localhost:8443/serviceworker/resources/fetch-access-control.php?mode=cors&credentials=same-origin&method=XXX&ACAOrigin=*&PACAOrigin=*&PACAMethods=XXX&PreflightTest=201 should be resolved expected "resolved" but got "rejected"
PASS executeTest-60 
PASS executeTest-61 
PASS executeTest-62 
FAIL executeTest-63 assert_equals: https://localhost:8443/serviceworker/resources/fetch-access-control.php?mode=cors&credentials=same-origin&method=XXX&headers=CUSTOM2&ACAOrigin=*&PACAOrigin=*&PACAMethods=XXX&PACAHeaders=x-servicEworker-u,x-servicEworker-ua,x-servicewOrker-test,x-sErviceworker-s,x-sErviceworker-v&PreflightTest=200 should be resolved expected "resolved" but got "rejected"
PASS executeTest-64 
FAIL executeTest-65 assert_equals: https://localhost:8443/serviceworker/resources/fetch-access-control.php?mode=cors&credentials=same-origin&method=XXX&headers=CUSTOM&ACAOrigin=*&PACAOrigin=*&PACAMethods=XXX&PACAHeaders=x-serviceworker-test&PACRMethod=XXX&PACRHeaders=x-serviceworker-test&PreflightTest=200 should be resolved expected "resolved" but got "rejected"
FAIL executeTest-66 assert_equals: https://localhost:8443/serviceworker/resources/fetch-access-control.php?mode=cors&credentials=same-origin&method=XXX&headers=SAFE&ACAOrigin=*&PACAOrigin=*&PACAMethods=XXX&PACRHeaders=missing&PACRMethod=XXX&PreflightTest=200 should be resolved expected "resolved" but got "rejected"
FAIL executeTest-67 assert_equals: https://localhost:8443/serviceworker/resources/fetch-access-control.php?mode=cors&credentials=same-origin&method=XXX&headers=CUSTOM2&ACAOrigin=*&PACAOrigin=*&PACAMethods=XXX&PACAHeaders=x-servicEworker-u,x-servicEworker-ua,x-servicewOrker-test,x-sErviceworker-s,x-sErviceworker-v&PACRMethod=XXX&PACRHeaders=x-serviceworker-s,x-serviceworker-test,x-serviceworker-u,x-serviceworker-ua,x-serviceworker-v&PreflightTest=200 should be resolved expected "resolved" but got "rejected"
Harness: the test ran to completion.
